Abort when download of ci.ri2.builds fails
While it's useful for a new repository to create this file, if it isn't
present, it is harmful when the download fails due to some network
error.

In this case the previous content of the file is dropped and replaced by
an empty file. This removes all previous packages from the list of installable
packages. This happened here:
https://ci.appveyor.com/project/larskanis/rubyinstaller2-packages/builds/23323519