Enable gcc -fstack-protector
This has two impacts:

It fixes missing references to __memcpy_chk in latest build by
implicit linking to libssp.

And it enables the stack-protector for building ruby and extensions.

The stack protector is enabled on most linux distros, so it's probably
a good idea to enable it on Windows as well. I verified that the
stack protector actually works: It stops the application with the
message "*** stack smashing detected ***:  terminated" after invalid
stack access.

See also https://github.com/msys2/MINGW-packages/issues/5803